+++ janitor/functions/__init__.py
@@ -81,7 +81,6 @@ from .utils import (
     col,
     get_columns,
     get_index_labels,
-    make_clean_names,
     patterns,
     unionize_dataframe_categories,
 )
@@ -130,7 +129,6 @@ __all__ = [
     "join_apply",
     "label_encode",
     "limit_column_characters",
-    "make_clean_names",
     "min_max_scale",
     "move",
     "pivot_longer",